+++ packages/vsf-storyblok-extension/hook.js
@@ -2,7 +2,7 @@ import { json, Router } from 'express'
 import request from 'request'
 import { promisify } from 'util'
 import { apiStatus } from '../../../lib/util'
-import { fullSync } from './fullSync'
+import { fullSync } from './sync'
 import { log } from './helpers'
 import protectRoute from './middleware/protectRoute'
 import { storyblokClient } from './storyblok'
